Isolate meshes by layer and remove old mesh-storage code (#6960)
* remove code related to mesh storage (i.e., STLs that are dragged into wk and would be stored on the server at some point)

* remove more mesh-storage related code

* isolate meshes by layer name

* fix initialization

* refactor mesh code into MeshController

* remove meshes property from public writes in annotation

* update snapshots

* update changelog

* rename mesh controller to segment mesh controller and move addLights code
