+++ frontend/javascripts/oxalis/model/reducers/reducer_helpers.ts
@@ -100,7 +100,6 @@ export function convertServerAnnotationToFrontendAnnotation(annotation: APIAnnot
     name,
     typ: annotationType,
     tracingStore,
-    meshes,
     owner,
     contributors,
     othersMayEdit,
@@ -116,7 +115,6 @@ export function convertServerAnnotationToFrontendAnnotation(annotation: APIAnnot
     name,
     annotationType,
     tracingStore,
-    meshes,
     owner,
     contributors,
     othersMayEdit,
