+++ frontend/javascripts/test/fixtures/skeletontracing_server_objects.ts
@@ -198,7 +198,6 @@ export const annotation: APIAnnotation = {
   },
   tags: ["ROI2017_wkw", "skeleton"],
   tracingTime: 0,
-  meshes: [],
   contributors: [],
   othersMayEdit: false,
 };
