+++ frontend/javascripts/test/fixtures/volumetracing_server_objects.ts
@@ -111,7 +111,6 @@ export const annotation: APIAnnotation = {
   },
   tags: ["ROI2017_wkw", "volume"],
   tracingTime: 0,
-  meshes: [],
   contributors: [],
   othersMayEdit: false,
 };
