+++ frontend/javascripts/test/snapshots/public-test/test-bundle/test/backend-snapshot-tests/annotations.e2e.js.md
@@ -26,7 +26,6 @@ Generated by [AVA](https://avajs.dev).
       description: '',
       formattedHash: '56fe9b',
       id: '570ba0092a7c0e980056fe9b',
-      meshes: [],
       modified: 0,
       name: '',
       organization: 'Organization_X',
@@ -146,7 +145,6 @@ Generated by [AVA](https://avajs.dev).
       description: '',
       formattedHash: '81c05d',
       id: '88135c192faeb34c0081c05d',
-      meshes: [],
       modified: 0,
       name: '',
       organization: 'Organization_X',
@@ -266,7 +264,6 @@ Generated by [AVA](https://avajs.dev).
       description: '',
       formattedHash: 'formattedHash',
       id: 'id',
-      meshes: [],
       messages: [
         {
           success: 'Task is finished',
@@ -438,7 +435,6 @@ Generated by [AVA](https://avajs.dev).
       description: '',
       formattedHash: 'formattedHash',
       id: 'id',
-      meshes: [],
       messages: [
         {
           success: 'Annotation was reopened',
@@ -610,7 +606,6 @@ Generated by [AVA](https://avajs.dev).
       description: '',
       formattedHash: 'formattedHash',
       id: 'id',
-      meshes: [],
       messages: [
         {
           success: 'Annotation is archived',
@@ -735,7 +730,6 @@ Generated by [AVA](https://avajs.dev).
       description: '',
       formattedHash: 'formattedHash',
       id: 'id',
-      meshes: [],
       messages: [
         {
           success: 'Annotation was reopened',
@@ -860,7 +854,6 @@ Generated by [AVA](https://avajs.dev).
       description: 'new description',
       formattedHash: 'formattedHash',
       id: 'id',
-      meshes: [],
       modified: 'modified',
       name: 'new name',
       organization: 'Organization_X',
@@ -980,7 +973,6 @@ Generated by [AVA](https://avajs.dev).
       description: '',
       formattedHash: 'formattedHash',
       id: 'id',
-      meshes: [],
       modified: 'modified',
       name: '',
       organization: 'Organization_X',
