+++ frontend/javascripts/test/snapshots/public-test/test-bundle/test/backend-snapshot-tests/tasks.e2e.js.md
@@ -420,7 +420,6 @@ Generated by [AVA](https://avajs.dev).
         description: '',
         formattedHash: '81c068',
         id: '58135c402faeb34e0081c068',
-        meshes: [],
         modified: 0,
         name: '',
         organization: 'Organization_X',
@@ -706,7 +705,6 @@ Generated by [AVA](https://avajs.dev).
       description: '',
       formattedHash: 'formattedHash',
       id: 'id',
-      meshes: [],
       modified: 'modified',
       name: '',
       organization: 'Organization_X',
