+++ frontend/javascripts/types/api_flow_types.ts
@@ -446,18 +446,6 @@ export function annotationToCompact(annotation: APIAnnotation): APIAnnotationCom
   };
 }
 
-export type LocalMeshMetaData = {
-  isVisible?: boolean;
-  isLoaded?: boolean;
-  isLoading?: boolean;
-};
-export type RemoteMeshMetaData = {
-  annotationId: string;
-  position: Vector3;
-  description: string;
-  id: string;
-};
-export type MeshMetaData = LocalMeshMetaData & RemoteMeshMetaData;
 export type AnnotationViewConfiguration = {
   layers: Record<
     string,
@@ -477,7 +465,6 @@ type APIAnnotationBase = APIAnnotationCompact & {
   readonly user?: APIUserBase;
   readonly contributors: APIUserBase[];
   readonly othersMayEdit: boolean;
-  readonly meshes: Array<MeshMetaData>;
 };
 export type APIAnnotation = APIAnnotationBase & {
   readonly task: APITask | null | undefined;
